Refactor Server.java
